# Conflicts:
#	resources/views/products/product_details.blade.php

+++ resources/views/admin/products.blade.php
@@ -2,7 +2,11 @@
 @section('content')
 
     <div class="admin-article">
+
         <div class="container">
+            <div class="article">
+                <a href="/admin"><button class="button float-left"><i class="fas fa-arrow-left pr-2 ml-0"></i>Terug</button></a>
+            </div>
             <table>
                 <thead>
                 <tr>
@@ -22,18 +26,21 @@
                         <td> {{$bike->omschrijving}}</td>
                         <td>
                             <div class="buttons">
-                                <a  href="{{$bike->id}}"><button class="button"><i class="fa fa-edit"></i></button></a>
-                                <a  href="delete/{{$bike->id}}"><button class="button button__delete"><i class="fa fa-trash"></i></button></a>
+                                <a  href="{{$bike->id}}" ><button class="button"><i class="fa fa-edit"></i></button></a>
+                                <a  href="delete/{{$bike->id}}"><button onclick="return confirm('Weet je het zeker?')" class="button button__delete"><i class="fa fa-trash"></i></button></a>
                             </div>
                         </td>
                     </tr>
                 @endforeach
 
 
+
                 </tbody>
+
             </table>
+            {{$bikes->links()}}
             <div class="article">
-                <button class="button"> Voeg product toe</button>
+                <a href="create"><button class="button"> Voeg product toe</button></a>
             </div>
         </div>
     </div>
